condition classes refactoring
